lambda/scraper/problem/requirements.txt
python-dotenv
boto3
requests
beautifulsoup4
pandas
aiohttp